Added a new app to serve the API
