+++ src/api/urls.py
@@ -2,7 +2,7 @@ from django.urls import path, include
 from rest_framework import routers
 from rest_framework.schemas import get_schema_view
 
-from base_app import api_views as views
+from api import views
 
 router = routers.DefaultRouter()
 router.register(r'file', views.FileViewSet)
@@ -10,6 +10,9 @@ router.register(r'transaction', views.TransactionViewSet)
 router.register(r'transaction-category', views.CategoryViewSet)
 router.register(r'currency', views.CurrencyViewSet)
 router.register(r'balance-details', views.BalanceDetailsViewSet)
+router.register(r'cash-transaction', views.CashTransactionViewSet)
+router.register(r'member', views.MemberViewSet)
+router.register(r'linked-member-transaction', views.LinkedTransactionViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
