+++ src/sports_accounting/urls.py
@@ -19,25 +19,16 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib.auth import views as auth_views
 
-from main import urls as main_urls
-from base_app import urls as base_app_urls
-
-from base_app import api_urls as api_base_app_urls
-from cash_module import api_urls as cash_module_urls
-from member_module import api_urls as member_module_urls
-
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/login/', auth_views.LoginView.as_view(), name='login'),
     path('accounts/logout/', auth_views.LogoutView.as_view(), name='logout'),
     # API (rest_framework)
+    path('api/', include('api.urls')),
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
-    path('api/', include(api_base_app_urls)),
-    path('api/', include(cash_module_urls)),
-    path('api/', include(member_module_urls)),
     # "Normal" views
-    path('', include(main_urls)),
-    path('mt940/', include(base_app_urls)),
+    path('', include('main.urls')),
+    path('mt940/', include('base_app.urls')),
 ]
 
 # In production, serve static files with a web server like Nginx or Apache
