+++ src/templates/main/index.html
@@ -1,7 +1,8 @@
 {% extends "main/base.html" %}
 
 {% block body %}
-	<h1>Welcome to this awesome Django Application!</h1>
+	<h1>Sports Accounting Main Page</h1>
+	<h1>This is a <i>Django Application!</i></h1>
 	{% if user.is_authenticated %}
 		<p>You are logged in as {{ user.username }}.</p>
 		<a href="{% url "logout" %}">Logout</a>
